chore(store): removed circular dependency preventing downstream package tests
