+++ packages/store/src/organizationSlice/constants.ts
@@ -11,9 +11,7 @@
 //    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
-import { Integration } from '@northern.tech/types/MenderTypes';
-
-import { EXTERNAL_PROVIDER, useradmApiUrl } from '../constants';
+import { useradmApiUrl } from '../constants';
 
 export const XML_METADATA_FORMAT = 'xml';
 export const JSON_METADATA_FORMAT = 'json';
@@ -62,21 +60,6 @@ export const auditlogTypes = {
 export const AUDIT_LOGS_TYPES = [auditlogTypes.artifact, auditlogTypes.deployment, auditlogTypes.device, auditlogTypes.user_access_token, auditlogTypes.user];
 export const SP_AUDIT_LOGS_TYPES = [auditlogTypes.user, auditlogTypes.tenant];
 
-export interface Webhook extends Integration {
-  provider: Integration.provider.WEBHOOK;
-}
-
-export const emptyWebhook: Webhook = {
-  description: '',
-  provider: Integration.provider.WEBHOOK,
-  credentials: {
-    type: EXTERNAL_PROVIDER.webhook.credentialsType,
-    [EXTERNAL_PROVIDER.webhook.credentialsType]: {
-      secret: '',
-      url: ''
-    }
-  }
-};
 export const TENANT_LIST_DEFAULT = {
   page: 1,
   perPage: 20
